requirements.txt
# Core LangChain dependencies
langchain>=0.1.0
langchain-community>=0.0.20

# Local LLM support
llama-cpp-python>=0.2.0

# Remote model APIs
openai>=1.0.0
anthropic>=0.7.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
langchain-google-genai>=2.0.0

# Database
sqlalchemy>=2.0.0

# Web requests
requests>=2.31.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0

# Monitoring (optional)
langsmith>=0.0.60

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Code quality
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Data validation
pydantic>=2.0.0

# UI (optional)
# Uncomment as needed:
streamlit>=1.28.0
# streamlit-chat>=0.1.0
# gradio>=3.40.0
# fastapi>=0.104.0
# uvicorn>=0.24.0

